Add task to list
