+++ src/style.css
@@ -101,8 +101,11 @@ form {
   align-items: center;
   padding:0px;
   background: white;
-  border-bottom: 1px solid rgba(0,0,0,0.08);
   width: 100%;
+}
+
+#add-item {
+  border-bottom: 1px solid rgba(0,0,0,0.08);
 }
 
 #add-item input[type=text] {
@@ -133,6 +136,10 @@ form {
   height: 18px;
   cursor: pointer;
   color: #333;
+}
+
+#submit-new-task:hover {
+  color: green;
 }
 
 .input-item {
